Convert DatabaseService to TypeScript
